Fix unlink/insertion/find race
It is only safe to retire a node when all of the higher entries in
the tower have been unlinked; otherwise a new Find invocation can
acquire a handle on the retired object post-deletion. The previous
algorithm assumed that reaching level-0 in the tower was sufficient
to prove that higher levels had been unlinked. This ignored the
possibility of a concurrent Erase call during Find, leading to
unlinking of only the lower portion of the tower.

This turns out to be difficult to solve via restart to the top of the
towers, because wiping out the next[i] entries is ambiguous for the
last node in the list (you can't tell whether it's been processed or
not, since it's already nullptr). There is also a race with
concurrent insertion/unlink.

This patch switches to a reference counting method that is used only
during node deletion. The reference count should exhibit low
contention and does not affect the read path.