Revert "Apple actually does support __thread"
This reverts commit bb2bdfed5677b508e4bd7266efdb24da6af5fb70.
Although OS X has support for __thread storage class, it interacts
badly with use of pthread_setspecific's on-exit callback to implement
dynamic thread local variables. In particular, it appears that the
TLS is no longer accessible upon invocation of the on-exit callback,
so we can't hook cleanup at that point. Contrast the situation on
Linux, where the TLS is available to the cleanup callback.